% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{indices_data}
\alias{indices_data}
\title{Example dataset of acoustic indices}
\format{
An object of class \code{data.frame} with 1635 rows and 94 columns.
}
\usage{
indices_data
}
\description{
A dataset of acoustic indices resulting from the function \code{\link[ANECO]{acoustic_indices}}, with the parameter \code{"noise.ind = TRUE"}
}
\keyword{datasets}
