% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAMCalib.R
\name{PAMCalib}
\alias{PAMCalib}
\title{Sound pressure level calibration}
\usage{
PAMCalib(
  ...,
  atype = "PSD",
  plottype = "Both",
  envi = "Air",
  calib = 0,
  ctype = "MF",
  Si = NULL,
  Mh = NULL,
  G = NULL,
  vADC = NULL,
  r = 50,
  N = Fs,
  winname = "Hann",
  lcut = Fs/N,
  hcut = Fs/2,
  timestring = "",
  outdir = dirname(fullfile),
  outwrite = 0,
  disppar = 0,
  welch = "",
  chunksize = "",
  linlog = "Log",
  tempWav = tempWav,
  ifile = ifile
)
}
\arguments{
\item{atype}{Ver Merchant et al. (2015)}

\item{plottype}{Ver Merchant et al. (2015)}

\item{envi}{Ver Merchant et al. (2015)}

\item{calib}{Ver Merchant et al. (2015)}

\item{ctype}{Ver Merchant et al. (2015)}

\item{Si}{Ver Merchant et al. (2015)}

\item{Mh}{Ver Merchant et al. (2015)}

\item{G}{Ver Merchant et al. (2015)}

\item{vADC}{Ver Merchant et al. (2015)}

\item{r}{Ver Merchant et al. (2015)}

\item{N}{Ver Merchant et al. (2015)}

\item{winname}{Ver Merchant et al. (2015)}

\item{lcut}{Ver Merchant et al. (2015)}

\item{hcut}{Ver Merchant et al. (2015)}

\item{timestring}{Ver Merchant et al. (2015)}

\item{outdir}{Ver Merchant et al. (2015)}

\item{outwrite}{Ver Merchant et al. (2015)}

\item{disppar}{Ver Merchant et al. (2015)}

\item{welch}{Ver Merchant et al. (2015)}

\item{chunksize}{Ver Merchant et al. (2015)}

\item{linlog}{Ver Merchant et al. (2015)}

\item{tempWav}{Character. Full path to the sound file}

\item{ifile}{Character. Sound file name}
}
\value{
A matrix with the values calibrated in decibels, each row is a temporal value of duration TimeRes and each column is a frequency value
}
\description{
It is an internal function of the \code{\link[ANECO]{ANECO}} package used in the function \code{\link[ANECO]{acoustic_indices}}. This function transform the intensity values of sound files into calibrated dB values according to technical parameters of the recorder device
}
\references{
\itemize{
\item{Merchant, N. D., Fristrup, K. M., Johnson, M. P., Tyack, P. L., Witt, M. J., Blondel, P., & Parks, S. E. (2015). Measuring acoustic habitats. Methods in Ecology and Evolution, 6(3), 257-265.}{}}
}
