% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundclassifier_pred.R
\name{soundclassifier_predict}
\alias{soundclassifier_predict}
\title{Prediction of heavy rain and cicadas noises
The function computes a random forest model to classify the minutes of a database into three categories: Clean, Rain and Cicada}
\usage{
soundclassifier_predict(data = NULL, mod = NULL)
}
\arguments{
\item{data}{A dataset generated by the function \code{\link[ANECO]{acoustic_indices}} (the option \code{"noise.ind = TRUE"} must have been included in the function \code{\link[ANECO]{acoustic_indices}})}

\item{mod}{A previously fitted random forest model to classify the data according to the selected indices.}
}
\value{
A dataset with the classification of each minute of the original dataset with its respective identification
}
\description{
Prediction of heavy rain and cicadas noises
The function computes a random forest model to classify the minutes of a database into three categories: Clean, Rain and Cicada
}
